Move Command constants into constants file
