+++ src/atlclients/basicInterceptor.ts
@@ -1,7 +1,7 @@
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from 'axios';
 import { commands, window } from 'vscode';
 
-import { Commands } from '../commands';
+import { Commands } from '../constants';
 import { Logger } from '../logger';
 import { AuthInfoState, DetailedSiteInfo } from './authInfo';
 import { AuthInterceptor } from './authInterceptor';
