+++ src/commands/bitbucket/rerunPipeline.ts
@@ -2,7 +2,7 @@ import { commands, window } from 'vscode';
 
 import { pipelineRerunEvent } from '../../analytics';
 import { clientForSite } from '../../bitbucket/bbUtils';
-import { Commands } from '../../commands';
+import { Commands } from '../../constants';
 import { Container } from '../../container';
 import { Logger } from '../../logger';
 import { Pipeline } from '../../pipelines/model';
