+++ src/extension.ts
@@ -8,9 +8,9 @@ import { startListening } from './atlclients/negotiate';
 import { BitbucketContext } from './bitbucket/bbContext';
 import { activate as activateCodebucket } from './codebucket/command/registerCommands';
 import { CommandContext, setCommandContext } from './commandContext';
-import { Commands, registerCommands } from './commands';
+import { registerCommands } from './commands';
 import { Configuration, configuration, IConfig } from './config/configuration';
-import { ExtensionId, GlobalStateVersionKey } from './constants';
+import { Commands, ExtensionId, GlobalStateVersionKey } from './constants';
 import { Container } from './container';
 import { registerAnalyticsClient, registerErrorReporting, unregisterErrorReporting } from './errorReporting';
 import { provideCodeLenses } from './jira/todoObserver';
