+++ src/views/pipelines/PipelinesExplorer.ts
@@ -2,9 +2,8 @@ import { commands, ConfigurationChangeEvent } from 'vscode';
 
 import { BitbucketContext } from '../../bitbucket/bbContext';
 import { CommandContext, setCommandContext } from '../../commandContext';
-import { Commands } from '../../commands';
 import { configuration } from '../../config/configuration';
-import { PipelinesTreeViewId } from '../../constants';
+import { Commands, PipelinesTreeViewId } from '../../constants';
 import { Container } from '../../container';
 import { BitbucketActivityMonitor } from '../BitbucketActivityMonitor';
 import { BitbucketExplorer } from '../BitbucketExplorer';
