+++ src/views/pipelines/PipelinesTree.ts
@@ -14,8 +14,8 @@ import {
 import { ProductBitbucket } from '../../atlclients/authInfo';
 import { clientForSite } from '../../bitbucket/bbUtils';
 import { WorkspaceRepo } from '../../bitbucket/model';
-import { Commands } from '../../commands';
 import { configuration } from '../../config/configuration';
+import { Commands } from '../../constants';
 import { Container } from '../../container';
 import { Pipeline } from '../../pipelines/model';
 import { Resources } from '../../resources';
