+++ src/views/pullRequestNodeDataProvider.ts
@@ -5,8 +5,8 @@ import { ProductBitbucket } from '../atlclients/authInfo';
 import { BitbucketContext } from '../bitbucket/bbContext';
 import { clientForSite } from '../bitbucket/bbUtils';
 import { PaginatedPullRequests, WorkspaceRepo } from '../bitbucket/model';
-import { Commands } from '../commands';
 import { configuration } from '../config/configuration';
+import { Commands } from '../constants';
 import { Container } from '../container';
 import { BaseTreeDataProvider } from './Explorer';
 import { GitContentProvider } from './gitContentProvider';
