+++ src/views/pullrequest/headerNode.ts
@@ -1,6 +1,6 @@
 import { TreeItem, TreeItemCollapsibleState } from 'vscode';
 
-import { Commands } from '../../commands';
+import { Commands } from '../../constants';
 import { Resources } from '../../resources';
 import { AbstractBaseNode } from '../nodes/abstractBaseNode';
