+++ src/views/pullrequest/prCommentController.ts
@@ -1,12 +1,13 @@
-import { fileCheckoutEvent, prCommentEvent, prTaskEvent } from 'src/analytics';
 import TurndownService from 'turndown';
 import { v4 } from 'uuid';
-import vscode, { commands, CommentThread, MarkdownString } from 'vscode';
+import * as vscode from 'vscode';
+import { commands, CommentThread, MarkdownString } from 'vscode';
 
+import { fileCheckoutEvent, prCommentEvent, prTaskEvent } from '../../analytics';
 import { BitbucketMentionsCompletionProvider } from '../../bitbucket/bbMentionsCompletionProvider';
 import { clientForSite } from '../../bitbucket/bbUtils';
 import { BitbucketSite, Comment, emptyTask, Task } from '../../bitbucket/model';
-import { Commands } from '../../commands';
+import { Commands } from '../../constants';
 import { Container } from '../../container';
 import { PullRequestNodeDataProvider } from '../pullRequestNodeDataProvider';
 import { PRFileDiffQueryParams } from './diffViewHelper';
@@ -790,7 +791,7 @@ export class PullRequestCommentController implements vscode.Disposable {
 
     provideComments(uri: vscode.Uri) {
         const { site, commentThreads, prHref, prId } = JSON.parse(uri.query) as PRFileDiffQueryParams;
-        (commentThreads || []).forEach(async (commentThread: Comment[]) => {
+        const promises = (commentThreads || []).map(async (commentThread: Comment[]) => {
             let range = new vscode.Range(0, 0, 0, 0);
             if (commentThread[0].inline!.from) {
                 range = new vscode.Range(commentThread[0].inline!.from! - 1, 0, commentThread[0].inline!.from! - 1, 0);
@@ -807,6 +808,8 @@ export class PullRequestCommentController implements vscode.Disposable {
                 await this.createOrUpdateThread(commentThread[0].id!, uri, range, comments);
             }
         });
+
+        return Promise.all(promises);
     }
 
     private async insertTasks(comments: PullRequestComment[]): Promise<vscode.Comment[]> {
