+++ src/views/pullrequest/pullRequestCreatedMonitor.ts
@@ -3,7 +3,7 @@ import * as vscode from 'vscode';
 
 import { BitbucketContext } from '../../bitbucket/bbContext';
 import { clientForSite } from '../../bitbucket/bbUtils';
-import { Commands } from '../../commands';
+import { Commands } from '../../constants';
 import { Logger } from '../../logger';
 import { BitbucketActivityMonitor } from '../BitbucketActivityMonitor';
